+++ src/DAISIE_DE_odeint.h
@@ -178,3 +178,84 @@ void integrate(const std::string& stepper_name,
 
 
 }   // namespace odeintcpp
+
+
+template <typename RaIt>
+inline double normalize_loglik(RaIt first, RaIt last) {
+  return 0.0;
+
+  const auto sabs = std::accumulate(first, last, 0.0,
+                                    [](const auto& s, const auto& x) {
+                                      return s + std::abs(x);
+                                    });
+  if (sabs <= 0.0) return 0.0;
+  const auto fact = 1.0 / sabs;
+  for (; first != last; ++first) *first *= fact;
+  return std::log(sabs);
+}
+
+template <typename ODE,
+          typename NORMALIZER>
+class Integrator {
+public:
+  using ode_type = ODE;
+
+  Integrator(std::unique_ptr<ode_type>&& od,
+             const std::string& method,
+             double atol,
+             double rtol) :
+    od_(std::move(od)),
+    method_(method),
+    atol_(atol),
+    rtol_(rtol)
+  {}
+
+  size_t size() const noexcept { return od_->size(); }
+
+  void operator()(std::vector<double>& state, double t0, double t1,
+                NORMALIZER& norm) const {
+    do_integrate(state, t0, t1, SECSSE_DEFAULT_DTF, norm);
+  }
+
+  void operator()(std::vector<double>& state, double t0, double t1) const {
+    odeintcpp::no_normalization no_norm;
+    do_integrate(state, t0, t1, SECSSE_DEFAULT_DTF, no_norm);
+  }
+
+  void operator()(std::vector<double> init_state,
+                std::vector<double>& times,
+                std::vector< std::vector<double>>* states_out) const {
+    odeintcpp::integrate(method_,
+                         od_.get(),
+                         init_state,
+                         times,
+                         SECSSE_DEFAULT_DTF,
+                         atol_,
+                         rtol_,
+                         states_out);
+  }
+
+private:
+  template <typename N>
+  void do_integrate(std::vector<double>& state,
+                    double t0,
+                    double t1,
+                    double dtf,
+                    N& norm) const {
+    odeintcpp::integrate(method_,
+                         od_.get(),
+                         &state,
+                         t0,
+                         t1,
+                         dtf * (t1 - t0),
+                         atol_,
+                         rtol_,
+                         norm);
+  }
+
+  std::unique_ptr<ODE> od_;
+  const std::string method_;
+  const double atol_;
+  const double rtol_;
+};                               // NOLINT [whitespace/indent]
+
